requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
chat-exporter==1.3.0.7
click==7.1.2
discord.py==1.7.2
DiscordUtils==1.2.9.2
emoji==1.2.0
Flask==1.1.2
grapheme==0.6.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==5.1.0
Pillow==8.2.0
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.4
Werkzeug==1.0.1
yarl==1.6.3
youtube-dl==2021.4.26